# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Mairie des Jeunes de Covè CMS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team:
    get:
      operationId: listTeam
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create team member (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete team member (admin)
      responses:
        "204":
          description: Deleted

  # ===== PROJECTS =====
...
components:
  schemas:
...
    TeamMember:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        role: { type: string }
        bio: { type: string }
        photoUrl: { type: string }
        email: { type: string }
        phone: { type: string }
        facebookUrl: { type: string }
        linkedinUrl: { type: string }
        sortOrder: { type: integer }
      required: [id, fullName, role, sortOrder]
    TeamMemberInput:
      type: object
      properties:
        fullName: { type: string }
        role: { type: string }
        bio: { type: string }
        photoUrl: { type: string }
        email: { type: string }
        phone: { type: string }
        facebookUrl: { type: string }
        linkedinUrl: { type: string }
        sortOrder: { type: integer }
      required: [fullName, role]
...